rewrote read tests to not run when loading xlsx raises an error

+++ tests/testthat/test_readfunctions.R
@@ -67,11 +67,15 @@ test_that("infer_names works as expected", {
 
 test_that("read_blocks reads data correctly", {
   #Make test blockcurves data
-  library(xlsx)
+  run_xlsx <- TRUE
+  tryCatch(
+    library(xlsx),
+    error = function(e) {run_xlsx <- FALSE}
+  )
   
   setwd(tempdir())
   dir.create("./test_blockcurves_data_csv/", showWarnings = F)
-  dir.create("./test_blockcurves_data_xlsx/", showWarnings = F)
+  if(run_xlsx) {dir.create("./test_blockcurves_data_xlsx/", showWarnings = F)}
   
   example_dfs_list <- rep(list(NA), 5)
   for (i in 1:length(example_dfs_list)) {
@@ -82,12 +86,14 @@ test_that("read_blocks reads data correctly", {
               paste("./test_blockcurves_data_csv/",
                     formatC(i, width = 3, flag = "0"), ".csv", sep = ""),
               row.names = TRUE)
-    write.xlsx(example_dfs_list[[i]],
-               file = paste("./test_blockcurves_data_xlsx/",
-                            formatC(i, width = 3, flag = "0"),
-                            ".xlsx", sep = ""),
-               sheetName = "Sheet1",
-               col.names = TRUE, row.names = TRUE, append = FALSE)
+    if(run_xlsx) {
+      write.xlsx(example_dfs_list[[i]],
+                 file = paste("./test_blockcurves_data_xlsx/",
+                              formatC(i, width = 3, flag = "0"),
+                              ".xlsx", sep = ""),
+                 sheetName = "Sheet1",
+                 col.names = TRUE, row.names = TRUE, append = FALSE)
+    }
   }
   
   #Read csv with all rows/cols specified, metadata included
@@ -111,42 +117,45 @@ test_that("read_blocks reads data correctly", {
   }
   expect_equal(my_blockcurves1, my_blockcurves1_expected)
   
-  my_blockcurves1_excel <- read_blocks(
+  my_blockcurves1_excelnumbers <- read_blocks(
     files = paste("./test_blockcurves_data_csv/",
                   list.files("./test_blockcurves_data_csv/"), sep = ""),
     startrow = 1, startcol = 1, endrow = 9, endcol = "M",
     metadata = list("type1" = c(9, "I"), "type2" = c("E", 8)))
-  expect_equal(my_blockcurves1_excel, my_blockcurves1_expected)
+  expect_equal(my_blockcurves1_excelnumbers, my_blockcurves1_expected)
   
   #Read xlsx with all rows/cols specified, metadata included
-  my_blockcurves2 <- read_blocks(
-    files = paste("./test_blockcurves_data_xlsx/",
-                  list.files("./test_blockcurves_data_xlsx/"), sep = ""),
-    startrow = 1, startcol = 1, endrow = 9, endcol = 13,
-    metadata = list("type1" = c(9, 9), "type2" = c(5, 8)))
-  my_blockcurves2_expected <- rep(list(NA), length(example_dfs_list))
-  for (i in 1:length(my_blockcurves2_expected)) {
-    my_blockcurves2_expected[[i]] <- 
-      list(data = example_dfs_list[[i]],
-           metadata = c(block_name = paste("test_blockcurves_data_xlsx/",
-                                           formatC(i, width = 3, flag = "0"),
-                                           sep = ""),
-                        type1 = example_dfs_list[[i]][8, 8],
-                        type2 = example_dfs_list[[i]][4, 7]))
-    row.names(my_blockcurves2_expected[[i]]$data) <-
-      as.character(row.names(my_blockcurves2_expected[[i]]$data))
+  if(run_xlsx) {
+    my_blockcurves2 <- read_blocks(
+      files = paste("./test_blockcurves_data_xlsx/",
+                    list.files("./test_blockcurves_data_xlsx/"), sep = ""),
+      startrow = 1, startcol = 1, endrow = 9, endcol = 13,
+      metadata = list("type1" = c(9, 9), "type2" = c(5, 8)))
+    my_blockcurves2_expected <- rep(list(NA), length(example_dfs_list))
+    for (i in 1:length(my_blockcurves2_expected)) {
+      my_blockcurves2_expected[[i]] <- 
+        list(data = example_dfs_list[[i]],
+             metadata = c(block_name = paste("test_blockcurves_data_xlsx/",
+                                             formatC(i, width = 3, flag = "0"),
+                                             sep = ""),
+                          type1 = example_dfs_list[[i]][8, 8],
+                          type2 = example_dfs_list[[i]][4, 7]))
+      row.names(my_blockcurves2_expected[[i]]$data) <-
+        as.character(row.names(my_blockcurves2_expected[[i]]$data))
+    }
+    expect_equal(my_blockcurves2, my_blockcurves2_expected)
+    
+    my_blockcurves2_excel <- read_blocks(
+      files = paste("./test_blockcurves_data_xlsx/",
+                    list.files("./test_blockcurves_data_xlsx/"), sep = ""),
+      startrow = 1, startcol = 1, endrow = 9, endcol = 13,
+      metadata = list("type1" = c("I", 9), "type2" = c(5, "H")))
+    expect_equal(my_blockcurves2_excel, my_blockcurves2_expected)
+    
+    unlink("./test_blockcurves_data_xlsx", recursive = TRUE)
   }
-  expect_equal(my_blockcurves2, my_blockcurves2_expected)
-  
-  my_blockcurves2_excel <- read_blocks(
-    files = paste("./test_blockcurves_data_xlsx/",
-                  list.files("./test_blockcurves_data_xlsx/"), sep = ""),
-    startrow = 1, startcol = 1, endrow = 9, endcol = 13,
-    metadata = list("type1" = c("I", 9), "type2" = c(5, "H")))
-  expect_equal(my_blockcurves2_excel, my_blockcurves2_expected)
   
   unlink("./test_blockcurves_data_csv", recursive = TRUE)
-  unlink("./test_blockcurves_data_xlsx", recursive = TRUE)
 })
 
 test_that("read_blocks reads data correctly for multiple blocks in one file", {
@@ -202,7 +211,11 @@ test_that("read_blocks reads data correctly for multiple blocks in one file", {
 
 test_that("read_wides works correctly", {
   #Make test data
-  library(xlsx)
+  run_xlsx <- TRUE
+  tryCatch(
+    library(xlsx),
+    error = function(e) {run_xlsx <- FALSE}
+  )
   
   setwd(tempdir())
   dir.create("./test_widecurves_data/", showWarnings = F)
@@ -219,50 +232,80 @@ test_that("read_wides works correctly", {
                        rnorm(100, sd = 0.5)))
   row.names(data) <- 2:101
   write.csv(data, "./test_widecurves_data/test.csv", row.names = FALSE)
-  write.xlsx(data, "./test_widecurves_data/test.xlsx",
-             sheetName = "Sheet1", col.names = TRUE, row.names = FALSE, 
-             append = FALSE)
+  if(run_xlsx) {
+    write.xlsx(data, "./test_widecurves_data/test.xlsx",
+               sheetName = "Sheet1", col.names = TRUE, row.names = FALSE, 
+               append = FALSE)
+  }
   
   #No names to col, no other settings specified
-  data_in <- read_wides(
-    files = c("./test_widecurves_data/test.csv",
-              "./test_widecurves_data/test.xlsx"),
-    names_to_col = NULL)
-  expect_equal(data_in,
-               list("test_widecurves_data/test" = data,
-                    "test_widecurves_data/test" = data))
+  if(!run_xlsx) {
+    data_in <- read_wides(
+      files = c("./test_widecurves_data/test.csv"),
+      names_to_col = NULL)
+    expect_equal(data_in,
+                 list("test_widecurves_data/test" = data))
+  } else {
+    data_in <- read_wides(
+      files = c("./test_widecurves_data/test.csv",
+                "./test_widecurves_data/test.xlsx"),
+      names_to_col = NULL)
+    expect_equal(data_in,
+                 list("test_widecurves_data/test" = data,
+                      "test_widecurves_data/test" = data))
+  }
   
-  
-  data_in2 <- read_wides(
-    files = c("./test_widecurves_data/test.csv",
-              "./test_widecurves_data/test.xlsx"),
-    names_to_col = "file")
   data2 <- cbind(data.frame(file = rep("test_widecurves_data/test", nrow(data))),
-                           data)
-  expect_equal(data_in2,
-               list("test_widecurves_data/test" = data2,
-                    "test_widecurves_data/test" = data2))
+                 data)
+  if(!run_xlsx) {
+    data_in2 <- read_wides(
+      files = c("./test_widecurves_data/test.csv"),
+      names_to_col = "file")
+    expect_equal(data_in2,
+                 list("test_widecurves_data/test" = data2))
+  } else {
+    data_in2 <- read_wides(
+      files = c("./test_widecurves_data/test.csv",
+                "./test_widecurves_data/test.xlsx"),
+      names_to_col = "file")
+    expect_equal(data_in2,
+                 list("test_widecurves_data/test" = data2,
+                      "test_widecurves_data/test" = data2))
+  }
   
-  data_in3 <- read_wides(
-    files = c("./test_widecurves_data/test.csv",
-              "./test_widecurves_data/test.xlsx"),
-    metadata = list("row5col5" = c(5, 5), "row12col2" = c(12, 2)))
   data3 <- cbind(file = data2[, 1],
                  #The row-col numbers are off bc of headers 
                  data.frame("row5col5" = rep(data[4, 5], nrow(data2)),
                             "row12col2" = rep(data[11, 2], nrow(data2))),
-                              data2[, -1])
-  expect_equal(data_in3,
-               list("test_widecurves_data/test" = data3,
-                    "test_widecurves_data/test" = data3))
-  
-  data_in3_excel <- read_wides(
-    files = c("./test_widecurves_data/test.csv",
-              "./test_widecurves_data/test.xlsx"),
-    metadata = list("row5col5" = c(5, "E"), "row12col2" = c(12, "B")))
-  expect_equal(data_in3_excel,
-               list("test_widecurves_data/test" = data3,
-                    "test_widecurves_data/test" = data3))
+                 data2[, -1])
+  if(!run_xlsx) {
+    data_in3 <- read_wides(
+      files = c("./test_widecurves_data/test.csv"),
+      metadata = list("row5col5" = c(5, 5), "row12col2" = c(12, 2)))
+    expect_equal(data_in3, list("test_widecurves_data/test" = data3))
+    
+    data_in3_excelnums <- read_wides(
+      files = c("./test_widecurves_data/test.csv"),
+      metadata = list("row5col5" = c(5, "E"), "row12col2" = c(12, "B")))
+    expect_equal(data_in3_excelnums,
+                 list("test_widecurves_data/test" = data3))
+  } else {
+    data_in3 <- read_wides(
+      files = c("./test_widecurves_data/test.csv",
+                "./test_widecurves_data/test.xlsx"),
+      metadata = list("row5col5" = c(5, 5), "row12col2" = c(12, 2)))
+    expect_equal(data_in3,
+                 list("test_widecurves_data/test" = data3,
+                      "test_widecurves_data/test" = data3))
+    
+    data_in3_excelnums <- read_wides(
+      files = c("./test_widecurves_data/test.csv",
+                "./test_widecurves_data/test.xlsx"),
+      metadata = list("row5col5" = c(5, "E"), "row12col2" = c(12, "B")))
+    expect_equal(data_in3_excelnums,
+                 list("test_widecurves_data/test" = data3,
+                      "test_widecurves_data/test" = data3))
+  }
   
   unlink("./test_widecurves_data", recursive = TRUE)
 })
